--- a/tests/kernels/test_layernorm.py
+++ b/tests/kernels/test_layernorm.py
@@ -1,8 +1,9 @@
+import time
+
 import torch
 import torch.nn as nn
-import time
-from vllm import layernorm_ops
 
+from vllm import layernorm_ops
 
 class RefRMSNorm(nn.Module):