I found that there is a kenel for writing subsequent optimizations in rmsnorm, and I tried to write a half-precision kernel for rms.
Below is the comparison data, I tested it on A100 80G, range 10000

### benchmark
num_tokens | hidden_size | Elapsed time (base) | Elapsed time (opt)
-- | -- | -- | --
2048 | 64 | 1.470 seconds | 1.594 seconds
2048 | 768 | 0.163 seconds | 0.066 seconds
2048 | 1024 | 0.171 seconds | 0.076 seconds
2048 | 5120 | 0.531 seconds | 0.280 seconds
4096 | 64 | 0.116 seconds | 0.090 seconds
4096 | 768 | 0.256 seconds | 0.073 seconds
4096 | 1024 | 0.311 seconds | 0.084 seconds
4096 | 5120 | 0.991 seconds | 0.497 seconds
10192 | 64 | 0.208 seconds | 0.158 seconds
10192 | 768 | 0.707 seconds | 0.178 seconds
10192 | 1024 | 0.851 seconds | 0.278 seconds
10192 | 5120 | 2.383 seconds | 1.164 seconds
### analyze
Except token=2048&&hidden=64 in the case, there can be a good speedup ratio, because in some cases the time is too small, and there may be fluctuations when you reproduce. The worse in the case of token=2048&&hidden=64 should be due to the fact that the number of threads started at this time is too small, but this situation should not exist in the real scene.





Since I used pytorch to register the kernel for the first time, the current method of registering the kernel may not be elegant. Doing benchmark can only comment out the half-precision kernel, and then recompile to test the time-consuming. If you have a better method, you can tell me I will modify it.