@sleepcoo 2023-06-28
> Hi @sleepcoo, Is the bug fixed now? We will add the code format checker later. 🙏 Could you wrap up this PR first?

I fixed it, you can review it when you have time, I also completed the fusion of softmax and div in sampler, which is twice the improvement compared to the original, I will submit another pr after you provide the code formatting tool